Refactor files to separate events from helpers
