+++ lib/events.js
@@ -1,7 +1,7 @@
 import { connect } from 'react-redux';
 
 export class Event {
-  constructor({ name, manager, listenTo = [] } = {}) {
+  constructor({ name, manager, useDataFrom, listenTo = [] } = {}) {
     if (arguments.length === 0) {
       throw new Error('An Event should not be initialized without parameters.');
     } else {
@@ -30,6 +30,7 @@ export class Event {
     this.name = name;
     this.manager = manager;
     this.listenTo = listenTo;
+    this.useDataFrom = useDataFrom;
     this.__UNSAFE_state = manager.initialState;
     this.reducerName = this._formatReducerName(this.name);
   }
@@ -122,8 +123,8 @@ export class Event {
 }
 
 export class HTTPEvent extends Event {
-  constructor({ name, manager, listenTo = [] } = {}) {
-    super({ name, manager, listenTo });
+  constructor({ name, manager, useDataFrom, listenTo = [] } = {}) {
+    super({ name, manager, useDataFrom, listenTo });
 
     delete this.reducerName;
 
@@ -176,38 +177,4 @@ export class HTTPEvent extends Event {
 
     return reducers;
   };
-}
-
-export function dispatchEvent({ event, store, data }) {
-  if (!event) {
-    throw new Error('You need to pass an event.');
-  } else if (typeof event.toRedux !== 'function') {
-    throw new Error(
-      'The event you passed needs to have a `toRedux` method. Are you sure you instantiated and passed the correct event?',
-    );
-  }
-
-  if (!store) {
-    throw new Error('You need to pass your redux store.');
-  } else if (typeof store.dispatch !== 'function') {
-    throw new Error(
-      'The store you passed does not have a `dispatch` method. Are you sure you passed the correct variable as the store?',
-    );
-  }
-
-  return store.dispatch(event.toRedux(data));
-}
-
-export function getCurrentStateFromEvent({ appState, event }) {
-  if (!event) {
-    throw new Error('You need to pass an event.');
-  }
-
-  if (!appState) {
-    throw new Error(
-      'You need to pass your app state. This is only available inside `shouldDispatch` methods or imported manually (not recommended).',
-    );
-  }
-
-  return appState[event.name];
 }
